Fix player rating calculation

+++ hw5/src/player.c
@@ -121,21 +121,23 @@ void player_post_result(PLAYER *player1, PLAYER *player2, int result) {
 		S2 = 0.5;
 	} else if(result == 1) {
 		/* player 1 wins */
-		S1 = 1.0;
-		S2 = 0.0;
+		S1 = 1;
+		S2 = 0;
 	} else if(result == 2) {
 		/* player 2 wins */
-		S1 = 0.0;
-		S2 = 1.0;
+		S1 = 0;
+		S2 = 1;
 	}
 	/* calculate score */
-	double E1 = (1.0 / (1.0 + pow(10, ((player2->rating - player1->rating) / 400))));
-	double E2 = (1.0 / (1.0 + pow(10, ((player1->rating - player2->rating) / 400))));
+	double E1 = 1.0 / (1 + pow(10, ((player2->rating - player1->rating) / 400.0)));
+	double E2 = 1.0 / (1 + pow(10, ((player1->rating - player2->rating) / 400.0)));
 	/* update ratings */
-	double R1 = (player1->rating + 32 * (S1 - E1));
-	double R2 = (player2->rating + 32 * (S2 - E2));
-	player1->rating = (int) R1;
-	player2->rating = (int) R2;
+	int R1 = player1->rating + (int)(32 * (S1 - E1));
+	int R2 = player2->rating + (int)(32 * (S2 - E2));
+	/* conserve total number of rating points */
+
+	player1->rating = R1;
+	player2->rating = R2;
 	/* unlock mutex */
 	if(pthread_mutex_unlock(&player2->mutex) != 0) {
 		debug("pthread_mutex_lock error");
